requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
appdirs==1.4.4
async-timeout==5.0.1
attrs==24.2.0
click==8.1.7
frozenlist==1.5.0
idna==3.10
iso8601==2.1.0
jedi==0.19.2
multidict==6.1.0
parso==0.8.4
prompt_toolkit==3.0.48
propcache==0.2.0
ptpython==3.0.29
pydantic==2.10.1
pydantic_core==2.27.1
Pygments==2.18.0
pypika-tortoise==0.2.2
pytz==2024.2
tomlkit==0.13.2
tortoise-cli==0.1.2
tortoise-orm==0.22.0
typing_extensions==4.12.2
wcwidth==0.2.13
yarl==1.18.0
